+++ Source/Interfaces/ConfigurationItemCollection.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace FreshAdvance\Dependency\Interfaces;
+
+/**
+ * Container of configuration items
+ */
+interface ConfigurationItemCollection
+{
+    /**
+     * @return array<ConfigurationItem>
+     */
+    public function getItems(): array;
+}